requirements.txt
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
dbt-core>=1.8.0
dbt-postgres>=1.8.0
dbt-labs/dbt-postgres>=1.8.0
python-telegram-bot>=20.7
python-telegram-bot-raw>=20.7
aiohttp>=3.9.1
pandas>=2.1.0
numpy>=1.26.0
python-dateutil>=2.8.2
pillow>=10.1.0
ultralytics>=8.0.20
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
